gigshield-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
apscheduler==3.10.4
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
python-multipart==0.0.12
requests==2.32.3
